Add Airport Seed

+++ JustInTimeCompany/Models/AircraftModel.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public int PassengerCapacity { get; set; }
-        public double Speed { get; set; }
+        public int Speed { get; set; }
 
         public ICollection<EngineInAircraft> Engines { get; init; }
     }
